+++ R/fixed.R
@@ -1,4 +1,4 @@
-#' @include distrib.R generics.R utility_functions.R moments.R cross_derivatives.R cross2_derivatives.R
+#' @include distrib.R generics.R utility_functions.R moments.R cross_derivatives.R cross2_derivatives.R cross_theta2_derivatives.R
 NULL
 
 #' @title S7 Class for Distributions With Fixed Parameters (Continuous)
@@ -173,6 +173,26 @@ for (.fixed_cls in list(FixedContinuousDistrib, FixedDiscreteDistrib)) {
       scale = "parameter", ...
     )
     res[distrib@params]
+  }
+
+  S7::method(distrib_grad_y_hess, .fixed_cls) <- function(distrib, y, theta,
+                                                          scale = c("parameter", "link"),
+                                                          ...) {
+    res <- distrib_grad_y_hess(distrib@parent_distrib, y,
+      fixed_full_theta(distrib, theta),
+      scale = "parameter", ...
+    )
+    res[hess_names(distrib@params)]
+  }
+
+  S7::method(distrib_hess_y_hess, .fixed_cls) <- function(distrib, y, theta,
+                                                          scale = c("parameter", "link"),
+                                                          ...) {
+    res <- distrib_hess_y_hess(distrib@parent_distrib, y,
+      fixed_full_theta(distrib, theta),
+      scale = "parameter", ...
+    )
+    res[hess_names(distrib@params)]
   }
 
   S7::method(distrib_gradient, .fixed_cls) <- function(distrib, y, theta,